CS50-PY/requirements.txt
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
emoji==2.10.1
fonttools==4.49.0
fpdf2==2.7.8
idna==3.6
inflect==7.0.0
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
pydantic==2.6.2
pydantic_core==2.16.3
pyfiglet==1.0.2
pytest==8.0.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
tabulate==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
validator-collection==1.5.0
